+++ tests/testthat/test-createStudy.R
@@ -32,6 +32,15 @@ test_that("Create a new v6 study", {
 })
 
 
-
+test_that("delete v8.1.0 study", {
+  path <- file.path(tempdir(), "tests_createStudy")
+  suppressWarnings(
+    opts <- createStudy(path, antares_version = "8.1.0")
+  )
+  properties <- antaresRead:::readIniFile(file.path(path, "study.antares"))
+  expect_identical(properties$antares$version, 810L)
+  expect_true(is_active_RES(opts))
+  deleteStudy(opts = simOptions())
+})