+++ NEWS.md
@@ -1,5 +1,4 @@
 
-
 # antaresEditObject 0.5.2
 
 ### Breaking changes (Antares v8.6) : 
@@ -7,6 +6,10 @@
 * Existing function `createBindingConstraint()` / `createBindingConstraintBulk()` has new arguments `group`
 * Existing function `editBindingConstraint()` has new arguments `group`
   * Integration of the scenario of the coupling constraints of second member (cf. Antares v8.6 changelog)
+
+### Breaking changes
+* `deleteStudy()` no longer requires user confirmation
+
 
 # antaresEditObject 0.5.1